refactor(runtime): move router support primitives
