+++ crates/atm-http-runtime/src/lib.rs
@@ -64,6 +64,7 @@ mod peer_connection_pool;
 mod peer_dial;
 mod peer_stream;
 mod private_staging;
+mod router_support;
 mod runtime_health;
 mod runtime_listener;
 mod runtime_maintenance;
